car-object-detection-torch/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "car_object_detection_torch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kaggle>=1.7.4.5",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]

[project.scripts]
car-object-detection = "car_object_detection_torch.main:run"

[tool.setuptools]
packages = ["car_object_detection_torch"]

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "pre-commit>=4.3.0",
    "ruff>=0.14.2",
]
